requirements.txt
Django>=3.2.9,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.7.5,<2.8.0

flake8>=4.0.1,<4.1.0